upd: декомпозиция страницы доски на компоненты
